requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
timm==0.9.12
kagglehub==0.2.3
scikit-learn==1.3.2
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.24.3
tqdm==4.66.1
